chore: Move last login notification to routes
