+++ web_ui/src/routes/app-routes.component.tsx
@@ -33,7 +33,6 @@ import { SignUp } from '../pages/sign-up/sign-up.component';
 import { TusUploadProvider } from '../providers/tus-upload-provider/tus-upload-provider.component';
 import { useWorkspaceIdentifier } from '../providers/workspaces-provider/use-workspace-identifier.hook';
 import { AccessDeniedDialog } from '../shared/components/access-denied-dialog/access-denied-dialog.component';
-import { LastLoginNotification } from '../shared/components/last-login-notification/last-login-notification';
 import { LicenseModal } from '../shared/components/license-modal/license-modal.component';
 import { IntelBrandedLoading } from '../shared/components/loading/intel-branded-loading.component';
 import { LOCAL_STORAGE_KEYS } from '../shared/local-storage-keys';
@@ -52,6 +51,7 @@ import { CameraScreenRoute } from './camera.route';
 import { MediaGalleryRoute } from './captured-media-gallery.route';
 import { IndexRoute } from './index.route';
 import { LandingPageLayout } from './landing-page.layout';
+import { LastLoginNotification } from './last-login-notification/last-login-notification';
 import { OrganizationsContext } from './organizations/organizations-context.component';
 import { UserManagementRoute } from './profile/index.route';
 import { AnnotatorRoute } from './projects/annotator.route';
